requirements.txt
gensim==4.1.2
matplotlib==3.3.4
nltk==3.6.1
numpy==1.19.5
pandas==1.2.4
scikit_learn==1.0.2
spacy==3.4
sumy==0.9.0

google-cloud-logging==2.7.0
google-cloud-speech==2.12.0
google-cloud-language==2.4.1
google-api-core==2.8.2
google-api-python-client==1.12.10
google-auth==2.9.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud==0.34.0
google-cloud-aiplatform==1.11.0
google-cloud-appengine-logging==1.1.2
google-cloud-audit-log==0.2.2
google-cloud-bigquery==2.33.0
google-cloud-bigquery-storage==2.12.0
google-cloud-core==2.3.1
google-cloud-scheduler==2.6.2
google-cloud-storage==1.44.0
google-crc32c==1.3.0
google-resumable-media==2.2.1
googleapis-common-protos==1.56.4

typing-inspect==0.8.0
typing_extensions==4.5.0